backend/requirements3-12-7.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-secret-manager
firebase_admin
requests
pandas
apache-airflow